+++ jwst/source_catalog/source_catalog.py
@@ -297,15 +297,6 @@ class Background:
         """
         Estimate the 2D background and background RMS noise in an image.
 
-        Parameters
-        ----------
-        box_size : int or array_like (int)
-            The box size along each axis.  If ``box_size`` is a scalar then
-            a square box of size ``box_size`` will be used.  If ``box_size``
-            has two elements, they should be in ``(ny, nx)`` order.
-
-        coverage_mask : bool ndarray
-
         Returns
         -------
         background : `photutils.background.Background2D`
@@ -995,7 +986,7 @@ class SourceCatalog:
         """
         The column names of the three concentration indices.
         """
-        return [f'CI_{self.aperture_ee[i]}_{self.aperture_ee[j]}'
+        return [f'CI_{self.aperture_ee[j]}_{self.aperture_ee[i]}'
                 for (i, j) in self._ci_ee_indices]
 
     @lazyproperty
@@ -1004,25 +995,28 @@ class SourceCatalog:
         The concentration indicies column descriptions.
         """
         return ['Concentration index calculated as '
-                f'{self.aperture_abmag_colnames[2*i]} - '
-                f'{self.aperture_abmag_colnames[2*j]}' for (i, j) in
+                f'({self.aperture_flux_colnames[2*j]} / '
+                f'{self.aperture_flux_colnames[2*i]})' for (i, j) in
                 self._ci_ee_indices]
 
     @lazyproperty
     def concentration_indices(self):
         """
-        A list of concentration indices, calculated as the difference of
-        AB magnitudes between:
+        A list of concentration indices, calculated as the flux
+        ratios of:
 
-            * the smallest and middle aperture radii/EE
-            * the middle and largest aperture radii/EE
-            * the smallest and largest aperture radii/EE
+            * the middle / smallest aperture radii/EE,
+              e.g., CI_50_30 = aper50_flux / aper30_flux
+            * the largest / middle aperture radii/EE,
+              e.g., CI_70_50 = aper70_flux / aper50_flux
+            * the largest / smallest aperture radii/EE,
+              e.g., CI_70_30 = aper70_flux / aper30_flux
         """
-        abmags = [(self.aperture_abmag_colnames[2*i],
-                   self.aperture_abmag_colnames[2*j]) for (i, j) in
+        fluxes = [(self.aperture_flux_colnames[2*j],
+                   self.aperture_flux_colnames[2*i]) for (i, j) in
                   self._ci_ee_indices]
-        return [getattr(self, mag1) - getattr(self, mag2)
-                for mag1, mag2 in abmags]
+        return [getattr(self, flux1).value / getattr(self, flux2).value
+                for flux1, flux2 in fluxes]
 
     def set_ci_properties(self):
         """
@@ -1200,20 +1194,6 @@ class SourceCatalog:
         return np.where(np.isnan(nn_abmag), nn_isomag, nn_abmag)
 
     @lazyproperty
-    def _not_star(self):
-        """
-        Whether the source is not a star.
-        """
-        if np.isnan(self.is_star[0]):
-            # TODO: remove this when ``is_star`` values are available.
-            # this array is all False:
-            is_star = np.zeros(len(self.id), dtype=bool)
-        else:
-            is_star = self.is_star
-
-        return np.logical_not(is_star)
-
-    @lazyproperty
     def aper_total_flux(self):
         """
         The aperture-corrected total flux for sources that are stars,
@@ -1222,8 +1202,7 @@ class SourceCatalog:
         idx = self.n_aper - 1  # apcorr for the largest EE (largest radius)
         flux = (self.aperture_params['aperture_corrections'][idx] *
                 getattr(self, self.aperture_flux_colnames[idx*2]))
-        flux[self._not_star] = np.nan
-
+        flux[~self.is_star] = np.nan
         return flux
 
     @lazyproperty
@@ -1235,8 +1214,7 @@ class SourceCatalog:
         idx = self.n_aper - 1  # apcorr for the largest EE (largest radius)
         flux_err = (self.aperture_params['aperture_corrections'][idx] *
                     getattr(self, self.aperture_flux_colnames[idx*2 + 1]))
-        flux_err[self._not_star] = np.nan
-
+        flux_err[~self.is_star] = np.nan
         return flux_err
 
     @lazyproperty